Hotfix/value from int to float

+++ app/src/main/java/com/example/finportfolio/fragments/assetdetails/AssetDetailsViewModel.kt
@@ -27,7 +27,7 @@ class AssetDetailsViewModel @Inject constructor(
         }
     }
 
-    fun addAssetToPortfolio(value: Int) {
+    fun addAssetToPortfolio(value: Float) {
         viewModelScope.launch {
             _assetModel.value?.let {
                 PortfolioAsset(
